Add first working PrepareCommitMessage-Hook
This introduces the first possibility to add the time since the last
git-commit to the commit message.

The configuration allows to use a certain fuzzines if needed to round up
to the next 5 or 10 minute intervall depending on the configuration.

The current implementation uses a configuration that rounds up to the
nex ten minute interval.

Time: 10h